requirements.txt
django-encrypted-model-fields
python-dateutil
asn1crypto==0.24.0
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.6.1
Django==2.1.7
django-cleanup==3.2.0
django-cors-headers==2.5.2
django-filter==2.1.0
django-microsoft-auth==2.2.0
django-oauth-toolkit==1.2.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
djangorestframework==3.9.2
idna==2.8
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
mysqlclient==1.4.2.post1
oauthlib==3.0.1
openapi-codec==1.3.2
pycparser==2.19
PyJWT==1.7.1
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
simplejson==3.16.0
six==1.12.0
uritemplate==3.0.0
urllib3==1.24.1
